Use a different timesource in recurrent timer.
Previously we used elapsedRealtimeNanos in recurrent timer which will
still go even if the system is in deep sleep. This causes all the
events "happened" during suspension to be replayed immediately
after the wake up, which causes a spam of messages. uptimeNanos
meanwhile, does not go if the system is in sleep, so we use that
instead.

Test: atest VehicleHalVehicleUtilsTest
Bug: 235262127
Change-Id: Ib67c2e2251af3231cefd875416d5bcb15953ba5e
Merged-In: Ib67c2e2251af3231cefd875416d5bcb15953ba5e
(cherry picked from commit 5e508737934feb595c841430f41b5edb1d1925d5)