Make new fields backward compatible.
ERROR: hardware/interfaces/radio/aidl/aidl_api/android.hardware.radio.network/3/android/hardware/radio/network/EmergencyRegResult.aidl:43.9-13: Field 'mcc' does not have a useful defau
lt in some backends. Please either provide a default value for this field or mark the field as @nullable. This value or a null value will be used automatically when an old version of t
his parcelable is sent to a process which understands a new version of this parcelable. In order to make sure your code continues to be backwards compatible, make sure the default or n
ull value does not cause a semantic change to this parcelable.
ERROR: hardware/interfaces/radio/aidl/aidl_api/android.hardware.radio.network/3/android/hardware/radio/network/EmergencyRegResult.aidl:44.9-13: Field 'mnc' does not have a useful defau
lt in some backends. Please either provide a default value for this field or mark the field as @nullable. This value or a null value will be used automatically when an old version of t
his parcelable is sent to a process which understands a new version of this parcelable. In order to make sure your code continues to be backwards compatible, make sure the default or n
ull value does not cause a semantic change to this parcelable.

Bug: 243966946
Test: time ./build/make/finalize-step-1.sh
Change-Id: I1a10dbee525d34407d44f5fa6c1159bd8d308265