Add @hide to the keymaster interfaces
This change allows the framework to depend on biometric AIDL types, for
example: `android.hardware.keymaster.HardwareAuthToken`.

The dependency can be introduced with a `-java-source` target, for
example: `:android.hardware.keymaster-V4-java-source`.

The "@hide" annotation informs Metalava that these types are not a
public API (not part of the SDK).

Bug: 218388821
Test: m android.hardware.keymaster-update-api
Change-Id: I0da555645f1548ad7124732236054f6e408c43e5