Remove lock for fakeVehicleHardware callbacks.
The callbacks are called very frequently, guarding them with the
main mLock will cause the mLock to become very hot and possibly
blocking other operations from owning the lock. Since the callback
is only set once by DefaultVehicleHal during initialization, it
is really not necessary to guard them with lock.

Test: manual run on gcar_emu
Bug: 255574557
Change-Id: Icc1f90b89578a27729ef8beae3a475966f72d318