audio VTS: Use the right device type in PcmOnlyConfigInputStreamTest
The fix in https://r.android.com/2322475 modifies the device
address, which becomes 'IN_DEFAULT' for the telephony case.
As a result, tests that verify capture position were not
skipped, and they do fail on certain platforms when there is
no active telephone call. The fix is to use the initial device
type which was passed to the test as a parameter, not the adjusted
one.

Bug: 263305254
Test: atest VtsHalAudioV7_0TargetTest
Test: atest VtsHalAudioV7_1TargetTest
Change-Id: Ied7134a1bd97e70dcc2400e0827810f9d71a6438