Fix hwasan error for CAN HAL
Hardware address sanitizer is complaining about parseConfigFile. I
suspect the issue is that protobuf is taking a pointer to an ifstream
rather than a reference, unique_ptr, or shared_ptr. I _think_ this
results in some sort of attempt to access the stream after it's closed.
In order to get around this, I moved ownership of the stream up one
level so that the stream stays open for longer.

Bug: 263769296
Test: canhalconfigurator-aidl doesn't crash on seahawk_hwasan-userdebug
Change-Id: I937d501a4759f0781304c518b518beaf8c6fed68