create staging/threadnetwork and add OWNERS
The `hardware/interfaces/staging/threadnetwork` folder contains the
unstable/unreleased `hardware/interfaces/threadnetwork` code which
are not supposed to be used in production. But vendors may start
verifying their hardware with the HAL interface.

This directory will be cleaned up when the stable Thread HAL interface
is added in `hardware/interfaces/threadnetwork` by version `V` or later.

Bug: 203492431
Change-Id: I0d6aee0666029fd17a94b3a46c51f43127db4759