Add a maximum MLO association link count
Maximum MLO association link count can be an input for MLO scoring
logic. If the candidate AP has less number of links to associate,
scroring logic can give a weightage to an AP with matching links as
station Assumption here is when the environment becomes better, the
AP with less score now can be a better choice.

Bug: 269336948
Test: m android.hardware.wifi-update-api
Change-Id: Iac0c82203c9e1a23637c757f5b92f5b213bb5da6