wifi: set join to true in p2p connect tests
After adding ag/21179155, auto join flag is
not set for every p2p connect command. The tests
are passing an unknown peer MAC address to connect(). To
pass the tests, the connect should trigger a scan and
try to find this unknown mac address. So set join flag
to true to run the same scan logic.

Bug: 268236829
Bug: 218608010
Test: atest VtsHalWifiSupplicantP2pIfaceTargetTest
Change-Id: Id5baa0f599d11a3563780aae15ff3cea0407060c