Add VTS to cover the APF VSR requirement since U
Since Android U, VSR 5.3.12 mentioned the Android Packet Filtering must
be supported by CHIPSETs and DEVICEs launching on Android 14 or higher.
This VTS was added to check if APF is supported properly.

Given the limitation of the VTS testing environment, the first version
of the test case only checks: if the APF is supported by WiFi chips with
version 4 and more than 1024 bytes of usable memory.

Bug: 259309511
Test: atest VtsHalWifiStaIfaceTargetTest
Change-Id: I3ee2a980a4440db232738a23c1633aaff3562d95