Update the setDtimMultiplier aidl API documentation.
Update the aidl API documentation to describe how setDtimMultiplier API
is supposed to work, i.e. we can call this API to set a maximum DTIM
multiplier value to use in driver, any larger value must not be accepted
and meanwhile it allows to set a specific multiplier by OEMs. Try to
make the description clear and avoid the ambiguity.

Bug: 266256943
Test: m
Change-Id: Ic7d2284a25e32d1b0752cf2bc832c069f646fcea