Re-enable radio satellite tests. radio to V3
This moves everything to radio-V3. What happened was satellite-V1 was
added with changes to the radio HAL in master. When radio-V2 was
frozen on goog/master, this showed up as radio-V2 having backwards
incompatible changes in goog/master. So I disabled the build files
there. Now that radio-V2 is frozen, we can re-enable the usage
of radio-V3 types in master (the VTS test was the only thing using
these new types).

In order to resolve this, I’ve moved all the radio libs over to V3.
Ideally, we could have moved one service over at a time, as
"satellite" support was needed. Though I’ve done this expecting other
radio libs to change in V3, this will mean we don’t expect more radio
.bp changes in Android V.

One alternative to this would have been to change the other radio libs
to import V3, but this is considered somewhat of a hack, and it would be
guaranteed to break any vendor-owned builds. With the current change,
vendor-owned build configs should still work as-is until they manually
add satellite.

Test: hal_implementation_test, vts_treble_vintf_vendor_test
Bug: 277379745

Change-Id: Iedc3097a9db58c0539e4aa61d089ee9d1986d51f