Use the default SAP interface name in VtsHalWifiHostapdV1_2TargetTest.
Test currently sets up a new interface
and uses that interface's name. On Pixel,
the new interface is wlan1, which does not
seem to support the OpenAccessPointWithoutAcs
configuration. However, the default SAP
interface wlan0 does support it.

The method to get the default interface name
is taken from this file before ag/19202002
was merged.

Bug: 270509029
Test: atest VtsHalWifiHostapdV1_2TargetTest
       (on a Bramble device running rvc-qpr3-release)
Change-Id: Ic3c516f452dfb54f155246073a9bca20dbecaa0f