thermal: Add support for Rust AIDL backend
Updates the VTS test to allow a NULL_POINTER exception when a null
pointer is passed as an argument. This matches the behavior of the Rust
libbinder_rs while maintaining backwards compatibility by still allowing
ILLEGAL_ARGUMENT exceptions in this case.

This support is necessary for a new emulated Thermal HAL for Cuttlefish
written in Rust.

Tested: VtsHalThermalTest in Cuttlefish
Bug: 286106286
Change-Id: Iccbebb416f31555da328f752d661660b77125ed3