[Thread] add "owner: google" for Thread HAL
This is for bypassing frozen HAL check on release branches.
Background:
We need to write code in the tethering module that depends on the
HAL interface and that code will be auto-merged to udc-mainline-prod
branch.

Merged-In is added to stop this CL from auto-merged to the main branch

Test: build
Bug: 296211911
Merged-In: I7e4bf260c03dd0500725fec068747e3b3d05820c
Change-Id: Id81adc9ff55649bf6d06cb2519d3e5805da15e5e