audio: Make IConfig.getSurroundSound default implementation more robust
In the case when there is a problem with the legacy APM XML file,
the converter provides a default surround sound config. However,
the default implementation of IConfig::getSurroundSoundConfig did
not take an advantage of that, and was returning an empty config,
which is not accepted by VTS.

Also, improve logging messages: clarify the situation when no readable
audio policy XML file found, and use outer functions name for lambdas.

Bug: 293978054
Test: atest VtsHalAudioCoreTargetTest
Change-Id: Iae069a0498009605ef5ededb9c9112efab08548a